requirements.txt
boto3
gunicorn
jmespath
libbam
libdna
libgeb
libgenomic
libgfb
libgff
libhttp
libhttpdna
libseq
numpy
psycopg2
s3fs
django
djangorestframework
graphene
graphene-django
